Computer-Vision/a-project-0/requirements.txt
pillow
matplotlib
opencv-python
numpy